Upgrade Rails 6.0 -> 6.1
Added active storage migration because it is standard and to gain more tests of the database adapter.
Installing active storage fits in this commit, because it is enabled by default in `config/application.rb` and would have to be disabled in there otherwise `rails app:update` creates an update migration that fails.
Using unofficial sqlserver adapter version with Rails 6.1 support